+++ tests/Feature/Request/RoaRequestTest.php
@@ -6,6 +6,7 @@ use AlibabaCloud\Client\AlibabaCloud;
 use AlibabaCloud\Client\Exception\ClientException;
 use AlibabaCloud\Client\Exception\ServerException;
 use AlibabaCloud\Client\Request\RoaRequest;
+use AlibabaCloud\Client\Tests\Mock\Services\Nlp\NlpRequest;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -37,20 +38,40 @@ class RoaRequestTest extends TestCase
                                         ->request();
             \assertNotEmpty($result->toArray());
         } catch (ClientException $e) {
-            self::assertContains(
-                $e->getErrorCode(),
-                [
-                    \ALIBABA_CLOUD_SERVER_UNREACHABLE,
-                ]
+            self::assertEquals(
+                \ALIBABA_CLOUD_SERVER_UNREACHABLE,
+                $e->getErrorCode()
             );
         } catch (ServerException $e) {
-            self::assertContains(
-                $e->getErrorMessage(),
-                [
-                    "cluster ($clusterId) not found in our records",
-                    'Specified access key is not found.',
-                ]
+            self::assertEquals(
+                "cluster ($clusterId) not found in our records",
+                $e->getErrorMessage()
             );
+        }
+    }
+
+    public function testRoaContent()
+    {
+        AlibabaCloud::accessKeyClient(
+            \getenv('NLP_ACCESS_KEY_ID'),
+            \getenv('NLP_ACCESS_KEY_SECRET')
+        )->name('content')
+                    ->regionId('cn-shanghai');
+
+        $request = new NlpRequest();
+        $request->pathParameter('Domain', 'general');
+        $request->jsonBody([
+                               'lang' => 'ZH',
+                               'text' => 'Iphone专用数据线',
+                           ]);
+
+        try {
+            $result = $request->client('content')->request();
+            self::assertEquals('Iphone', $result['data'][0]['word']);
+        } catch (ServerException $e) {
+            $this->assertEquals($e->getErrorCode(), 'InvalidApi.NotPurchase');
+        } catch (ClientException $e) {
+            self::assertStringStartsWith('cURL error', $e->getErrorMessage());
         }
     }
 }