+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.0.1 - 2019-01-09
+
+* Distinguish credential error.
+* Add endpoints for NLS.
+* Add not found product tip.
+
 ## 1.0.0 - 2019-01-07
 
 * Initial release of the Alibaba Cloud Client for PHP Version 1.0.0 on Packagist See <https://github.com/aliyun/openapi-sdk-php-client> for more information.